refactor starts

+++ src/main/java/com/pyding/deathlyhallows/rituals/rites/ElderRite.java
@@ -1,6 +1,5 @@
 package com.pyding.deathlyhallows.rituals.rites;
 
-import com.emoniph.witchery.ritual.RitualStep;
 import net.minecraft.entity.item.EntityItem;
 import net.minecraft.util.AxisAlignedBB;
 import net.minecraft.world.World;
@@ -22,14 +21,14 @@ public abstract class ElderRite {
 		double midX = 0.5 + (double)x;
 		double midZ = 0.5 + (double)z;
 		ArrayList<EntityItem> resultList = new ArrayList();
-		AxisAlignedBB bounds = AxisAlignedBB.getBoundingBox(midX - (double)radius, (double)y, midZ - (double)radius, midX + (double)radius, 1.0 + (double)y, midZ + (double)radius);
+		AxisAlignedBB bounds = AxisAlignedBB.getBoundingBox(midX - (double)radius, y, midZ - (double)radius, midX + (double)radius, 1.0 + (double)y, midZ + (double)radius);
 		List items = world.getEntitiesWithinAABB(EntityItem.class, bounds);
 		Iterator i$ = items.iterator();
 
 		while(i$.hasNext()) {
 			Object obj = i$.next();
 			EntityItem entity = (EntityItem)obj;
-			if (entity.getDistanceSq(midX, (double)y, midZ) <= (double)RADIUS_SQ) {
+			if(entity.getDistanceSq(midX, y, midZ) <= (double)RADIUS_SQ) {
 				resultList.add(entity);
 			}
 		}
